proj/hps_accel: inline CFU functions
Inline CFU functions to allow compiler to optimize switch statements
away. This reduces HPS evaluation time from 430s to 120s.
